remove this harmless warning message
I encountered the following when generating the site using this plugin:

```
Error: [WARNING] This document format requires a nonempty <title> element.
  Please specify either 'title' or 'pagetitle' in the metadata.
  Falling back to 'Untitled'
```

What happened is that HTML5, and so does Pandoc, requires a title to be specified for the output.
The warning message is then caught on
[this line](https://github.com/wzpan/hexo-renderer-pandoc/blob/dd837a149e39045b64577a226c5e77c7c39927d9/index.js#L72).
It then propagates to
[this line](https://github.com/wzpan/hexo-renderer-pandoc/blob/dd837a149e39045b64577a226c5e77c7c39927d9/index.js#L82)
and triggered 
[this line](https://github.com/wzpan/hexo-renderer-pandoc/blob/dd837a149e39045b64577a226c5e77c7c39927d9/index.js#L84)
to issue the error.

The issue is that this warning is harmless and should not become an error. So I propose this solution.

I originally wanted to match the result on
[this line](https://github.com/wzpan/hexo-renderer-pandoc/blob/dd837a149e39045b64577a226c5e77c7c39927d9/index.js#L72)
to the warning message and filter it out, but turned out that one warning message can span multiple data events, i.e., it can be returned in pieces.
So I decided to wait until the very end to remove this particular warning message.